+++ common/CHANGELOG.md
@@ -3,3 +3,4 @@
 ### Fixed
 
 - [Explosion Proof Jacket] incorrect damage reduction calculation.
+- [Nelumbo] [DarkNelumbo] fix not checking trinket for lava/water walking therefor allowing the player to walk in these fluids freely